Migrate to .slnx and restructure into src/ and tests/ directories
- Migrate RequesterMini.sln to RequesterMini.slnx (XML-based format)
- Add solution folders: /src/ and /tests/
- Move library projects (AppLogger, CurlExporter, JsonFileStore) into src/
- Move test projects into tests/
- Move main app source into src/RequesterMini/
- Update all project references and remove stale DefaultItemExcludes
