--- RequesterMini.csproj
@@ -1,51 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net10.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <BuiltInComInteropSupport>true</BuiltInComInteropSupport>
-    <ApplicationManifest>app.manifest</ApplicationManifest>
-    <AvaloniaUseCompiledBindingsByDefault>true</AvaloniaUseCompiledBindingsByDefault>
-  <PublishTrimmed>true</PublishTrimmed>
-  <PublishSingleFile>true</PublishSingleFile>
-  <JsonSerializerIsReflectionEnabledByDefault>false</JsonSerializerIsReflectionEnabledByDefault>
-  <DefaultItemExcludes>$(DefaultItemExcludes);CurlExporter\**;CurlExporter.Tests\**;JsonFileStore\**;JsonFileStore.Tests\**;AppLogger\**;AppLogger.Tests\**</DefaultItemExcludes>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <AvaloniaResource Include="Assets\**" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include=".github\copilot-instructions.md" />
-    <None Include="AppLogger.Tests\AppLogger.Tests.csproj" />
-    <None Include="AppLogger\AppLogger.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Avalonia" Version="11.3.12" />
-    <PackageReference Include="Avalonia.Desktop" Version="11.3.12" />
-    <PackageReference Include="Avalonia.Themes.Fluent" Version="11.3.12" />
-    <PackageReference Include="Avalonia.Fonts.Inter" Version="11.3.12" />
-    <!--Condition below is needed to remove Avalonia.Diagnostics package from build output in Release configuration.-->
-    <PackageReference Condition="'$(Configuration)' == 'Debug'" Include="Avalonia.Diagnostics" Version="11.3.12" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="10.0.3" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="10.0.3" />
-    <PackageReference Include="OneOf" Version="3.0.271" />
-    <PackageReference Include="ReactiveUI.Avalonia" Version="11.3.8" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <TrimmerRootAssembly Include="ReactiveUI" />
-    <TrimmerRootAssembly Include="ReactiveUI.Avalonia" />
-    <TrimmerRootAssembly Include="DynamicData" />
-    <TrimmerRootAssembly Include="Splat" />
-    <TrimmerRootAssembly Include="System.Reactive" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="AppLogger\AppLogger.csproj" />
-    <ProjectReference Include="CurlExporter\CurlExporter.csproj" />
-    <ProjectReference Include="JsonFileStore\JsonFileStore.csproj" />
-  </ItemGroup>
-</Project>
