--- ViewModels/ViewModelBase.cs
@@ -1,7 +1,0 @@
-﻿using ReactiveUI;
-
-namespace RequesterMini.ViewModels;
-
-public class ViewModelBase : ReactiveObject
-{
-}
