+++ src/AppLogger/AppLogger.csproj
