+++ src/CurlExporter/CurlCommandBuilder.cs
