+++ src/JsonFileStore/JsonFileStore.csproj
