+++ src/RequesterMini/Assets/avalonia-logo.ico
