+++ src/RequesterMini/Constants/HttpConstants.cs
