+++ src/RequesterMini/Models/RequestFailure.cs
