+++ src/RequesterMini/ViewModels/JsonVisualizerWindowViewModel.cs
