+++ src/RequesterMini/ViewModels/ViewModelBase.cs
@@ -1,0 +1,7 @@
+﻿using ReactiveUI;
+
+namespace RequesterMini.ViewModels;
+
+public class ViewModelBase : ReactiveObject
+{
+}
