+++ src/RequesterMini/Views/JsonVisualizerWindow.axaml.cs
