+++ tests/AppLogger.Tests/LoggerTests.cs
