+++ tests/CurlExporter.Tests/CurlExporter.Tests.csproj
@@ -15,7 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\CurlExporter\CurlExporter.csproj" />
+    <ProjectReference Include="..\..\src\CurlExporter\CurlExporter.csproj" />
   </ItemGroup>
 
   <ItemGroup>
